Move photo picker to its own package under ui
