+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoPickerPreviewActivity.java
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.posts.photopicker;
+package org.wordpress.android.ui.photopicker;
 
 import android.content.pm.ActivityInfo;
 import android.graphics.Bitmap;
