+++ WordPress/src/main/res/layout/new_edit_post_activity.xml
@@ -26,6 +26,6 @@
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
         android:visibility="gone"
-        tools:context="org.wordpress.android.ui.posts.photopicker.PhotoPickerFragment"
+        tools:context=".ui.photopicker.PhotoPickerFragment"
         tools:visibility="visible" />
 </RelativeLayout>
